Dockerfile
Uses the official Haskell docker image[1]. Since `stack.yaml` specifies `lts-16.24` as the resolver, the image tag is the matching GHC version, `8.8.4`.

Sets a working directory and copies in the file necessary to build the project. Then copies the rest of the project before specifying the `stack` entrypoint with the `run` command.

`docker build -t meatbar .` takes several minutes to `stack build`, which might be related to the `stack.yaml` `resolver` pulling in more than is necessary. There might be a more specific base image with that LTS.

`docker run -p 8080:8080 meatbar` run the built image and bind the port as needed, but appears to hang because there is no output indicating the executable is "ready." However, going to `localhost:8080` shows expected logging of server requests.

1: https://hub.docker.com/_/haskell